# lesson-14/docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "University of GO.IT",
    "version": "1.0.0",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of University of GO.IT"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:9090"
    }
  ],
  "tags": [
    {
      "name": "Students",
      "description": "Operation with students"
    }
  ],
  "paths": {
    "/students/{id}": {
      "get": {
        "tags": ["Students"],
        "summary": "Get student by id",
        "operationId": "getStudent",
        "description": "Get student by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67eec4c00c73bfc979baedbc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get student by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Student fetched successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/student"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "404": {
            "description": "Student not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 404
                    },
                    "message": {
                      "type": "string",
                      "example": "Student not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/students": {
      "post": {
        "tags": ["Students"],
        "summary": "Create student",
        "operationId": "createStudent",
        "description": "Create student",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["name", "gender", "year"],
                "properties": {
                  "name": {
                    "description": "Student's name",
                    "type": "string",
                    "example": "John Doe"
                  },
                  "gender": {
                    "description": "Student's gender",
                    "type": "string",
                    "enum": ["male", "female"],
                    "example": "male"
                  },
                  "year": {
                    "description": "Student's year",
                    "type": "integer",
                    "example": 1998
                  },
                  "onDuty": {
                    "description": "Whether is student on duty",
                    "type": "boolean",
                    "example": false,
                    "default": false
                  },
                  "avatar": {
                    "description": "Student's avatar",
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created a student",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status", "message", "data"],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully created a student"
                    },
                    "data": {
                      "$ref": "#/components/schemas/student"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "student": {
        "type": "object",
        "required": [
          "_id",
          "name",
          "gender",
          "year",
          "userId",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Student id",
            "example": "67eec4c00c73bfc979baedbc"
          },
          "name": {
            "type": "string",
            "description": "Student name",
            "example": "John Doe"
          },
          "gender": {
            "type": "string",
            "description": "Student gender",
            "enum": ["male", "female"],
            "example": "male"
          },
          "year": {
            "type": "number",
            "description": "Student born year",
            "example": 1998
          },
          "onDuty": {
            "type": "boolean",
            "description": "Student duty",
            "example": false,
            "default": false
          },
          "userId": {
            "type": "string",
            "description": "Student's teacher id",
            "example": "67eec78c0f9979f4b1081e52"
          },
          "avatar": {
            "type": "string",
            "description": "Student avatar url",
            "example": "https://res.cloudinary.com/dha4xvgfi/image/upload/v1743104601/snunhxlqvnd67budfeja.webp",
            "default": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Student created date",
            "example": "2025-03-27T19:43:21.686+00:00"
          },
          "updateAt": {
            "type": "string",
            "format": "date-time",
            "description": "Student updated date",
            "example": "2025-03-27T19:43:21.686+00:00"
          }
        }
      }
    },
    "responses": {
      "401": {
        "description": "Unauthorized Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message", "data"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "Unauthorized Error"
                },
                "data": {
                  "type": "object",
                  "required": ["message"],
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Access token not provided"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["status", "message"],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 500
                },
                "message": {
                  "type": "string",
                  "example": "Internal Server Error"
                }
              }
            }
          }
        }
      }
    }
  }
}